+++ packages/db-mongodb/src/queries/sanitizeQueryValue.ts
@@ -77,6 +77,7 @@ export const sanitizeQueryValue = ({
     // Object equality requires the value to be the first key in the object that is being queried.
     if (
       operator === 'equals' &&
+      formattedValue &&
       typeof formattedValue === 'object' &&
       formattedValue.value &&
       formattedValue.relationTo
@@ -156,6 +157,23 @@ export const sanitizeQueryValue = ({
 
   if (operator === 'exists') {
     formattedValue = formattedValue === 'true' || formattedValue === true
+
+    // Clearable fields
+    if (['relationship', 'select', 'upload'].includes(field.type)) {
+      if (formattedValue) {
+        return {
+          rawQuery: {
+            $and: [{ [path]: { $exists: true } }, { [path]: { $ne: null } }],
+          },
+        }
+      } else {
+        return {
+          rawQuery: {
+            $or: [{ [path]: { $exists: false } }, { [path]: { $eq: null } }],
+          },
+        }
+      }
+    }
   }
 
   return { operator: formattedOperator, val: formattedValue }