+++ packages/payload/src/collections/operations/local/find.ts
@@ -5,9 +5,7 @@ import type { Payload } from '../../../payload'
 import type { Document, Where } from '../../../types'
 
 import { APIError } from '../../../errors'
-import { setRequestContext } from '../../../express/setRequestContext'
-import { i18nInit } from '../../../translations/init'
-import { getDataLoader } from '../../dataloader'
+import { createLocalReq } from '../../../utilities/createLocalReq'
 import find from '../find'
 
 export type Options<T extends keyof GeneratedTypes['collections']> = {
@@ -39,32 +37,20 @@ export default async function findLocal<T extends keyof GeneratedTypes['collecti
 ): Promise<PaginatedDocs<GeneratedTypes['collections'][T]>> {
   const {
     collection: collectionSlug,
-    context,
     currentDepth,
     depth,
     disableErrors,
     draft = false,
-    fallbackLocale: fallbackLocaleArg = options?.req?.fallbackLocale,
     limit,
-    locale: localeArg = null,
     overrideAccess = true,
     page,
     pagination = true,
-    req = {} as PayloadRequest,
     showHiddenFields,
     sort,
-    user,
     where,
   } = options
-  setRequestContext(req, context)
 
   const collection = payload.collections[collectionSlug]
-  const localizationConfig = payload?.config?.localization
-  const defaultLocale = localizationConfig ? localizationConfig.defaultLocale : null
-  const locale = localeArg || req.locale || defaultLocale
-  const fallbackLocale = localizationConfig
-    ? localizationConfig.locales.find(({ code }) => locale === code)?.fallbackLocale
-    : null
 
   if (!collection) {
     throw new APIError(
@@ -72,17 +58,7 @@ export default async function findLocal<T extends keyof GeneratedTypes['collecti
     )
   }
 
-  req.payloadAPI = req.payloadAPI || 'local'
-  req.locale = locale
-  req.fallbackLocale =
-    typeof fallbackLocaleArg !== 'undefined' ? fallbackLocaleArg : fallbackLocale || defaultLocale
-  req.i18n = i18nInit(payload.config.i18n)
-  req.payload = payload
-
-  if (!req.t) req.t = req.i18n.t
-  if (!req.payloadDataLoader) req.payloadDataLoader = getDataLoader(req)
-
-  if (typeof user !== 'undefined') req.user = user
+  const req = createLocalReq(options, payload)
 
   return find<GeneratedTypes['collections'][T]>({
     collection,