+++ packages/payload/src/fields/hooks/beforeChange/promise.ts
@@ -88,6 +88,8 @@ export const promise = async ({
           global,
           operation,
           originalDoc: doc,
+          previousSiblingDoc: siblingDoc,
+          previousValue: siblingDoc[field.name],
           req,
           siblingData,
           value: siblingData[field.name],