+++ templates/ecommerce/src/payload/collections/Orders/hooks/clearUserCart.ts
@@ -6,7 +6,7 @@ export const clearUserCart: AfterChangeHook<Order> = async ({ doc, req, operatio
   const { payload } = req
 
   if (operation === 'create' && doc.orderedBy) {
-    const orderedBy = typeof doc.orderedBy === 'string' ? doc.orderedBy : doc.orderedBy.id
+    const orderedBy = typeof doc.orderedBy === 'object' ? doc.orderedBy.id : doc.orderedBy
 
     const user = await payload.findByID({
       collection: 'users',